move reminder to ext + proper reminder time task
